pyproject.toml
[project]
name = "lyrathon-wooloolies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic==0.75.0",
    "apify-client==1.5.0",
    "fastapi==0.104.1",
    "httpx==0.25.1",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "sqlalchemy==2.0.23",
    "uvicorn[standard]==0.24.0",
    "websockets==12.0",
]
